AS7-753 The start of EJB support in the console
This code makes it possible to register two different presenters for the
same subsystem. In the case of EJB3 there need to be two different
presenters, one for stateless session beans and another for MDBs, there
will be some overlap but each of them will have a specific focus. These
changes make it possible. The actual Session/MDB views don't contain
much over and above a few labels yet.