AS7-753 Start or rework of EJBs based on Stan's GUI framework
The layout is not yet completely up to scratch and needs revisiting, but
this adds EJB editors for the EJB3 subsystem defaults, pools and Timer
Service based on Stan's GUI framework.
There's more work to do but I thought it might be useful to merge this.
Note that the previous ejb console code is also removed in this commit.