AS7-1410 Runtime OSGi Management through console
Added Framework panel to show/edit start level.
Added Sorting to Bundle runtime view on ID, BSN and state.
Moved SingleEntityToDmrBridgeImpl to framework as it is a generally
useful class for cases where a single entity is to be edited through the
framework.
Added Refresh to Bundle runtime view.
Added proper handling of OSGi Bundle fragments.
Added 'auto activate' a setting that asks the user whether it should
automatically activate the OSGi subsystem if the user opens the runtime
view.
Modified the EntityToDmrBridgeImpl so that it include runtime
attributes.
Renamed AbstractEntityView.getPluralEntityName() to
getEntityDisplayName().