AS7-1335 Initial support for security domains.
The security domains have a lot of details associated with them. They
are currently places in tabs, however only the 'Attributes' and
'Authorization' currently have actual content.
Also in this commit I introduced a doCancel() method on the
EntityToDmrBridge which is consistent with the doEdit() and doSave() ones
already there. A doCancel() makes it possible to integrate non-framework
panels with other ones provided by the framework in such a way that
cancel works for these too.