Fixed typo - chose should probably be choose
While adding a deployment, I noticed some text in the dialog said "Chose
a file".  From there, I searched for other occurences of "chose" and
where it seemed that "choose" was intended, I made the change.  This was
not an automated find/replace, so if there are errors in there, it's
totally my fault.